#93 docs: signIn -> signUp 수정
