+++ src/page/signUp/index/TermPage.style.ts
