+++ src/page/signUp/index/component/TermArea/TermArea.style.ts
