+++ src/page/signUp/index/component/TermsAgreeButton/TermsAgreeButton.style.ts
