+++ src/page/signUp/info/constant/index.ts
